requirements.txt
praw==7.8.1
requests
moviepy==1.0.3
transformers==4.30.0
librosa==0.9.2
soundfile==0.12.1
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
moviepy==1.0.3
mutagen
lameenc
